Fixes APLO-99, APLO-100, and APLO-101 Adding more options to `SUBSCRIBE` so you can implement Kafka style durable pub/sub over a queue.
Support a `browser-end:false` option on a `SUBSCRIBE` frame to allow a Queue browser to wait for new messages instead of closing the subscription.
Support a `include-seq` and `from-seq` option on the `SUBSCRIBE` frame to finely control from where in a queue a subscription starts from.
Support Kafka style durable pub/sub.

git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-apollo/trunk@1210707 13f79535-47bb-0310-9956-ffa450edef68